requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.3
asyncssh==2.14.2
atpublic==4.0
attrs==23.2.0
billiard==4.2.0
cachetools==5.3.2
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
configobj==5.0.8
contourpy==1.2.0
cryptography==42.0.2
cycler==0.12.1
dictdiffer==0.9.0
diskcache==5.6.3
distlib==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
docopt==0.6.2
dpath==2.1.6
dulwich==0.21.7
entrypoints==0.4
filelock==3.13.1
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.12.2
funcy==2.0
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.16.1
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.62.0
grandalf==0.8
grpcio==1.60.0
gto==1.6.2
httplib2==0.22.0
hydra-core==1.3.2
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
iterative-telemetry==0.0.8
Jinja2==3.1.3
kiwisolver==1.4.5
kombu==5.3.5
librosa==0.9.1
lightning==2.1.3
lightning-utilities==0.10.1
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.8.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
numpy
oauth2client==4.1.3
oauthlib==3.2.2
omegaconf==2.3.0
onnxruntime
orjson==3.9.12
packaging==23.2
pandas==2.1.4
pathspec==0.12.1
pillow==10.2.0
pipreqs==0.4.13
platformdirs==3.11.0
prompt-toolkit==3.0.43
protobuf==4.25.1
psutil==5.9.8
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser
pydantic==2.6.0
pydantic_core==2.16.1
pydot==2.0.0
PyDrive2==1.19.0
pygit2==1.14.0
Pygments==2.17.2
pygtrie==2.5.0
pyOpenSSL==24.0.0
pyparsing==3.1.1
pystoi==0.4.1
python-dateutil==2.8.2
pytorch-lightning==2.1.3
pytz==2023.4
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
scmrepo==2.1.1
scipy==1.12.0
semver==3.0.2
sentry-sdk==1.40.0
setproctitle==1.3.3
shortuuid==1.0.11
shtab==1.6.5
six==1.16.0
smmap==5.0.1
soundfile
speechmos==0.0.1.1
sqltrie==0.11.0
sympy==1.12
tabulate==0.9.0
trimesh
tomlkit==0.12.3
torch==2.2.2
torchaudio==2.2.2
torchmetrics==1.2.1
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.0
vine==5.1.0
virtualenv==20.25.0
wandb
wheel==0.41.2
wcwidth==0.2.13
Werkzeug==3.0.1
yarg==0.1.9
yarl==1.9.4
zc.lockfile==3.0.post1
zipp==3.17.0
